Type Extension Support for Schema Stitching (#700)
