+++ src/Core/Types/Configuration/Contracts/ISchemaConfigurationExtension.cs
@@ -1,0 +1,9 @@
+using System;
+
+namespace HotChocolate.Configuration
+{
+    public interface ISchemaConfigurationExtension
+    {
+        ISchemaConfiguration OnBeforeBuild(Action<ISchemaBuilder> build);
+    }
+}
