+++ src/Stitching/Stitching.Tests/Middleware/DelegateToRemoteSchemaMiddlewareTests.cs
@@ -28,15 +28,14 @@ using HotChocolate.Types.Relay;
 namespace HotChocolate.Stitching
 {
     public class DelegateToRemoteSchemaMiddlewareTests
-        : IClassFixture<TestServerFactory>
+        : StitchingTestBase
     {
         public DelegateToRemoteSchemaMiddlewareTests(
             TestServerFactory testServerFactory)
+            : base(testServerFactory)
         {
-            TestServerFactory = testServerFactory;
         }
 
-        private TestServerFactory TestServerFactory { get; set; }
 
         [Fact]
         public async Task ExecuteStitchingQueryWithInlineFragment()
@@ -807,96 +806,6 @@ namespace HotChocolate.Stitching
         }
 
         [Fact]
-        public async Task ConnectionLost()
-        {
-            // arrange
-            var connections = new Dictionary<string, HttpClient>();
-            IHttpClientFactory clientFactory = CreateRemoteSchemas(connections);
-
-            var serviceCollection = new ServiceCollection();
-            serviceCollection.AddSingleton(clientFactory);
-            serviceCollection.AddStitchedSchema(builder =>
-                builder.AddSchemaFromHttp("contract")
-                    .AddSchemaFromHttp("customer")
-                    .RenameType("CreateCustomerInput", "CreateCustomerInput2")
-                    .AddExtensionsFromString(
-                        FileResource.Open("StitchingExtensions.graphql"))
-                    .AddSchemaConfiguration(c =>
-                        c.RegisterType<PaginationAmountType>())
-                    .AddExecutionConfiguration(b =>
-                    {
-                        b.AddErrorFilter(error =>
-                        {
-                            if (error.Exception is Exception ex)
-                            {
-                                return ErrorBuilder.FromError(error)
-                                    .ClearExtensions()
-                                    .SetMessage(ex.GetType().FullName)
-                                    .SetException(null)
-                                    .Build();
-                            };
-                            return error;
-                        });
-                    }));
-
-            IServiceProvider services =
-                serviceCollection.BuildServiceProvider();
-
-            IQueryExecutor executor = services
-                .GetRequiredService<IQueryExecutor>();
-            IExecutionResult result = null;
-
-            using (IServiceScope scope = services.CreateScope())
-            {
-                var request = new QueryRequest(@"
-                    mutation {
-                        createCustomer(input: { name: ""a"" })
-                        {
-                            customer {
-                                name
-                                contracts {
-                                    id
-                                }
-                            }
-                        }
-                    }");
-                request.Services = scope.ServiceProvider;
-
-                result = await executor.ExecuteAsync(request);
-            }
-
-            var client = new HttpClient
-            {
-                BaseAddress = new Uri("http://127.0.0.1")
-            }; ;
-            connections["contract"] = client;
-            connections["customer"] = client;
-
-            // act
-            using (IServiceScope scope = services.CreateScope())
-            {
-                var request = new QueryRequest(@"
-                    mutation {
-                        createCustomer(input: { name: ""a"" })
-                        {
-                            customer {
-                                name
-                                contracts {
-                                    id
-                                }
-                            }
-                        }
-                    }");
-                request.Services = scope.ServiceProvider;
-
-                result = await executor.ExecuteAsync(request);
-            }
-
-            // assert
-            Snapshot.Match(result);
-        }
-
-        [Fact]
         public async Task StitchedMutationWithRenamedInputType()
         {
             // arrange
@@ -1203,41 +1112,6 @@ namespace HotChocolate.Stitching
                 requestBuilder.SetServices(scope.ServiceProvider);
                 return await executor.ExecuteAsync(requestBuilder.Create());
             }
-        }
-
-        private IHttpClientFactory CreateRemoteSchemas()
-        {
-            return CreateRemoteSchemas(new Dictionary<string, HttpClient>());
-        }
-
-        private IHttpClientFactory CreateRemoteSchemas(
-            Dictionary<string, HttpClient> connections)
-        {
-            TestServer server_contracts = TestServerFactory.Create(
-                ContractSchemaFactory.ConfigureSchema,
-                ContractSchemaFactory.ConfigureServices,
-                new QueryMiddlewareOptions());
-
-            TestServer server_customers = TestServerFactory.Create(
-                CustomerSchemaFactory.ConfigureSchema,
-                CustomerSchemaFactory.ConfigureServices,
-                new QueryMiddlewareOptions());
-
-            connections["contract"] = server_contracts.CreateClient();
-            connections["customer"] = server_customers.CreateClient();
-
-            var httpClientFactory = new Mock<IHttpClientFactory>();
-            httpClientFactory.Setup(t => t.CreateClient(It.IsAny<string>()))
-                .Returns(new Func<string, HttpClient>(n =>
-                {
-                    if (connections.ContainsKey(n))
-                    {
-                        return connections[n];
-                    }
-
-                    throw new Exception();
-                }));
-            return httpClientFactory.Object;
         }
     }
 }
